Split info command
