+++ cmd/util-db/info/deleted_account.go
@@ -1,0 +1,56 @@
+package info
+
+import (
+	"fmt"
+	"github.com/0xsoniclabs/aida/cmd/util-db/flags"
+	"github.com/0xsoniclabs/aida/logger"
+	"github.com/0xsoniclabs/aida/utils"
+	"github.com/0xsoniclabs/substate/db"
+	"github.com/urfave/cli/v2"
+	"strings"
+)
+
+var printDeletedAccountsCommand = cli.Command{
+	Action:    printDeletedAccountsAction,
+	Name:      "del-acc",
+	Usage:     "Prints deletion info about an account in AidaDb.",
+	ArgsUsage: "<firstBlockNum>, <lastBlockNum>",
+	Flags: []cli.Flag{
+		&utils.AidaDbFlag,
+		&logger.LogLevelFlag,
+		&flags.Account,
+	},
+}
+
+// printDeletedAccountsAction for given deleted account in AidaDb
+func printDeletedAccountsAction(ctx *cli.Context) error {
+	cfg, argErr := utils.NewConfig(ctx, utils.BlockRangeArgs)
+	if argErr != nil {
+		return argErr
+	}
+
+	log := logger.NewLogger(cfg.LogLevel, "AidaDb-Deleted-Account-Info")
+
+	db, err := db.NewReadOnlyDestroyedAccountDB(cfg.DeletionDb)
+	if err != nil {
+		return err
+	}
+
+	accounts, err := db.GetAccountsDestroyedInRange(cfg.First, cfg.Last)
+	if err != nil {
+		return fmt.Errorf("cannot Get all destroyed accounts; %w", err)
+	}
+
+	wantedAcc := ctx.String(flags.Account.Name)
+
+	for _, acc := range accounts {
+		if strings.Compare(acc.String(), wantedAcc) == 0 {
+			log.Noticef("Account %v, got deleted in %v - %v", wantedAcc, cfg.First, cfg.Last)
+			return nil
+		}
+	}
+
+	log.Warningf("Account %v, didn't get deleted in %v - %v", cfg.First, cfg.Last)
+
+	return nil
+}
