+++ cmd/util-db/info/info.go
@@ -1,0 +1,38 @@
+// Copyright 2024 Fantom Foundation
+// This file is part of Aida Testing Infrastructure for Sonic
+//
+// Aida is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Lesser General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Aida is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU Lesser General Public License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public License
+// along with Aida. If not, see <http://www.gnu.org/licenses/>.
+
+package info
+
+import (
+	"github.com/urfave/cli/v2"
+)
+
+var Command = cli.Command{
+	Name:  "info",
+	Usage: "Prints information about AidaDb",
+	Subcommands: []*cli.Command{
+		&printDeletedAccountsCommand,
+		&printCountCommand,
+		&printRangeCommand,
+		&printStateHashCommand,
+		&printBlockHashCommand,
+		&printExceptionsCommand,
+		&printTableHashCommand,
+		&printPrefixHashCommand,
+		&dumpSubstateCommand,
+		&dbHashCommand,
+	},
+}
